// src/app/core/services/car/models/car.interface.ts
export interface ApiCarI{
    model : string;
    manufacturer : string;
    vehicle: string
    fuel: string;
    type: string;
    image: string;
    id: string;
    price: string
}

export interface CarI{
    model : string;
    manufacturer : string;
    vehicle: string
    fuel: string;
    type: string;
    image: string;
    id: string;
    price: string
}

export interface CarRequestBody {
    model : string;
    manufacturer : string;
    vehicle: string
    fuel: string;
    type: string;
    image: string;
    price: string
}